+++ components/urlParser.js
@@ -4,6 +4,7 @@ import qs from 'query-string';
 const getValue = (key, value) => {
 	const parsed = qs.parse(window.location.search);
 
+	// add `|| value` to save the config file's contents into localStorage in addition to the query-string args
 	if (parsed[key]) {
 		if (typeof window !== 'undefined') {
 			window.localStorage.setItem(key, parsed[key]);
@@ -21,15 +22,15 @@ const getValue = (key, value) => {
 const getConfig = (config) => {
 	let newConfig = {};
 	const parsed = qs.parse(window.location.search);
-	for (let key in config) {
+	for (const key in config) {
 		const value = config[key];
 
 		newConfig[key] = getValue(key, value);
 	}
 
-	const stringified = qs.stringify({...parsed, ...newConfig});
+	const stringified = qs.stringify(parsed);
 	if (typeof window !== 'undefined') {
-		window.history.replaceState('', '', `/?${stringified}`);
+		window.history.replaceState('', '', (stringified ? `/?${stringified}` : ''));
 	}
 
 	return newConfig;
